+++ packages/decimal/__tests__/decimal.test.ts
@@ -548,18 +548,6 @@ test('shift right rounding modes', () => {
   expect(shr('-1.6', 1, m)).toEqual(parse('-2'));
   expect(shr('-2.5', 1, m)).toEqual(parse('-2'));
   expect(shr('-5.5', 1, m)).toEqual(parse('-6'));
-
-  m = '05up';
-  expect(shr('5.5', 1, m)).toEqual(parse('6'));
-  expect(shr('2.5', 1, m)).toEqual(parse('2'));
-  expect(shr('1.6', 1, m)).toEqual(parse('1'));
-  expect(shr('1.1', 1, m)).toEqual(parse('1'));
-  expect(shr('1.0', 1, m)).toEqual(parse('1'));
-  expect(shr('-1.0', 1, m)).toEqual(parse('-1'));
-  expect(shr('-1.1', 1, m)).toEqual(parse('-1'));
-  expect(shr('-1.6', 1, m)).toEqual(parse('-1'));
-  expect(shr('-2.5', 1, m)).toEqual(parse('-2'));
-  expect(shr('-5.5', 1, m)).toEqual(parse('-6'));
 });
 
 test('trailing zeros count', () => {