+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [UNRELEASED]
+### Fixed/Changed
+ - Removed unused timezone aliases already covered in tzdb / `@phensley/timezone`
+
 ## [0.11.4]
 ### Fixed/Changed
  - Improved coefficient adjustment in scientific formatting