+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
-## [UNRELEASED]
+## [0.12.1]
 ### Fixed/Changed
  - Removed unused timezone aliases already covered in tzdb / `@phensley/timezone`
  - Changed types used in number parts representations:
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
    - Scientific exponent marker is `'exp'`.
  - Removed non-ICU rounding mode `'05up'`
  - Removed redundant rounding mode `'truncate'`. Use `'down'`.
+ - Ensure timezone id resolution maps back to CLDR stable id for exemplar city resolution
 
 ## [0.11.4]
 ### Fixed/Changed
@@ -77,6 +78,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 #### Fixed/Changed
 - Converted `CLDR` to an interface
 
+[0.12.1]: https://github.com/phensley/cldr-engine/compare/v0.11.4...v0.12.1
 [0.11.4]: https://github.com/phensley/cldr-engine/compare/v0.11.3...v0.11.4
 [0.11.3]: https://github.com/phensley/cldr-engine/compare/v0.10.2...v0.11.3
 [0.10.2]: https://github.com/phensley/cldr-engine/compare/v0.9.1...v0.10.2